move server code in separate file
